refactor: split main.rs into cmd modules and util.rs
- Extract 11 subcommand handlers into crates/lkr-cli/src/cmd/
- Extract shared helpers (confirm, clipboard, unlock) into util.rs
- main.rs reduced from 1319 to 516 lines (Cli + dispatch + error handling + tests)
- All 91 existing tests pass, no behavioral changes
