+++ crates/lkr-cli/src/cmd/mod.rs
@@ -1,0 +1,11 @@
+pub(crate) mod exec;
+pub(crate) mod r#gen;
+pub(crate) mod get;
+pub(crate) mod harden;
+pub(crate) mod init;
+pub(crate) mod list;
+pub(crate) mod lock;
+pub(crate) mod migrate;
+pub(crate) mod rm;
+pub(crate) mod set;
+pub(crate) mod usage;
